не легка задача була)
