+++ models/contacts.json
@@ -76,5 +76,17 @@
     "name": "Jack",
     "email": "example@mail.com",
     "phone": "1110000"
+  },
+  {
+    "id": "22K_c9JTKQV4QUucVWOO9",
+    "name": "Robert",
+    "phone": "1110000",
+    "email": "example@mail.com"
+  },
+  {
+    "id": "rqfgDVuc9YKWl5IJ0s8gd",
+    "name": "RobDimaert",
+    "email": "example@mail.com",
+    "phone": "1110000"
   }
 ]
