+++ utils/validation/validationSchema.js
@@ -1,0 +1,17 @@
+const Joi = require("joi");
+
+const validationSchema = () => {
+  const schema = Joi.object({
+    name: Joi.string().alphanum().min(3).max(30).required(),
+    email: Joi.string()
+      .email({
+        minDomainSegments: 2,
+        tlds: { allow: ["com", "net"] },
+      })
+      .required(),
+    phone: Joi.number().integer().required(),
+  });
+  return schema;
+};
+
+module.exports = { validationSchema };
